+++ run_freeze.py
@@ -1,0 +1,4 @@
+from setuputil import bump_requirements
+
+if __name__ == "__main__":
+    bump_requirements()